% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tomatrix.R
\name{tomatrix}
\alias{tomatrix}
\title{Put totalized vector into a matrix table}
\usage{
tomatrix(data, row, namelength = 7)
}
\arguments{
\item{data}{totalized vector.}

\item{row}{grouping index vector to be arranged by row. the others are to be arranged by column.}

\item{namelength}{optional. dimnames would be abbrevieated to this length.}
}
\description{
Put totalized vector into a matrix table.
}
\examples{
totalize(esoph, "ncases", c("agegp", "alcgp", "tobgp")) |> tomatrix(c(1, 2))

}
