% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totalize-package.R
\docType{package}
\name{totalize-package}
\alias{totalize-package}
\title{totalize: Aggregate values by groups and make their total then put into a matrix table}
\description{
To count or aggregate values by groups can be done by ordinary R methods, while 'totalize' makes their total additionaly.
It is useful especially when grouping data by 2 or more variables combination.
The return value is provided in arranged matrix representation.
}
\author{
\strong{Maintainer}: Ichiro Murata \email{ichiro21.murata@gmail.com}

}
