% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totalize.R
\name{totalize}
\alias{totalize}
\title{Aggregate values of each group and total group}
\usage{
totalize(data, target, by, FUN = sum, drop = FALSE, ...)
}
\arguments{
\item{data}{a data frame.}

\item{target}{a single column name.}

\item{by}{a character vector of column names for grouping.}

\item{FUN}{function to be applied for \code{target}.}

\item{drop}{whether to drop or retain unused factors in \code{by}.}

\item{...}{further arguments passed to \code{FUN}.}
}
\value{
A numeric vector consists of the results of \code{FUN} applied to each group and sum of the groups.
For \code{drop = FALSE}, NA will be created for unused grouping. This helps to make a matrix.

The returned value has 3 attributes. \code{name} is concatenated grouping names, \code{value} is the column name to which \code{FUN} has applied,
and \code{groups} is a list of grouping names with "all" category added.
}
\description{
Aggregate values of each group and total group.
}
\details{
\code{FUN} is applied not only to each of the groups but also a sum of the groups.
}
\examples{
totalize(CO2, "uptake", c("Type", "Treatment", "conc"), FUN=mean) |> tomatrix(3)

totalize(esoph, "ncases", c("agegp", "alcgp", "tobgp")) |> tomatrix(c(1, 2))

}
